+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Moved `RobotModel`, `ToolModel` and `Configuration` to the top-level package of this library.
 * Moved `URDF` parsing from core into this library.
 * Changed urdf parsing to use `compas.colors.Color` instead of a robot-specific color class.
+* Changed `ToolModel.link_name` to `ToolModel.connected_to`.
+* Fixed data serialization for `RobotModel` and `ToolModel`.
 
 ### Removed